requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
pydantic==2.5.0
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# AI/ML
groq==0.4.1
sentence-transformers==2.2.2
tenacity==8.2.3

# Vector database
qdrant-client==1.7.0

# Web and visualization
requests==2.31.0
plotly==5.17.0

# Logging and utilities
python-multipart==0.0.6
httpx==0.25.2 